test: Add E2E test for first install
Add E2E test to ensure a window opens upon first install.

This test needs to use a production-like build rather than a standard E2E
build, so the existing "vault decryptor" job was repurposed to be more
generically for E2E tests using a production-like build.

A global function `reloadExtension` is added to `stateHooks` for use by this
test. This function had to be enabled for production builds because the test
uses a production build.